v4l2viewer: Updated viewer recipe and removed patches
The recipe for the V4L2Viewer now contains a custom do_install function, which does not use cmake install anymore. Because of this the cmake install patch could be removed.
The latest version of the viewer detects the supported io-methods (USERPTR or MMAP) automaticly. Therefore the patch for setting the io-method to MMAP was removed.