Added extra device trees and configuration tool for 2 camera streaming
The streaming with GenICam for CSI2 on both ports in parallel on the imx8mpevk was not working, because the limitions of the isi 0 are different when two camera are used in parallel. Therefore, the original device tree is split into three seperate device trees and a configuration tool is added.
The first and default device tree allows streaming with two cameras in parallel with GenICam for CSI2 and v4l2. In v4l2 mode the width of both cameras is limitied to 2048.
The other device tree allow streaming with one port either CSI0 or CSI1. Both device trees use isi0 so it is possible to connected cameras with a width up to 4096 to both CSI ports.