The imx8mp media driver is now always loaded after the alvium driver
Previously the imx8mp media driver and the alvium driver were loaded independently. If the probe of the alvium driver took to long, it cloud happen that the camera was not detected correctly. This happened especially when using two cameras, because with two cameras the probe of the alvium driver takes longer. To fix this a configuration file is added, which declares a soft dependency between the alvium and imx8mp media driver. By adding the dependency it is ensured that the imx8mp media driver is always loaded after the alvium driver.